Оптимиз кода, добав файл queries для запросов
