+++ inc/function.php
@@ -11,29 +11,6 @@ $user_name = 'Михаил Лебедев';
 */
 $response_code = ''; 
 
-/* Общее подключение к БД */
-
-function getConn() {
-    $conn = mysqli_connect("localhost", "root", "", "yeticave");
-    mysqli_set_charset($conn, "utf8"); // первым делом кодировка
-    if (!$conn) {
-        print("Ошибка: Невозможно подключиться к MySQL " . mysqli_connect_error()); 
-        // die('Ошибка при подключении: ' . mysqli_connect_error()); // вариант 2.
-    }
-    return $conn;
-}
-
-/* Общий запрос категорий из БД таблицы без защиты от sql-инъекции, тк нет переменных */
-
-function getCategories($conn) {
-    $sql = 'SELECT symbol, name FROM categories'; 
-    $result = mysqli_query($conn, $sql);
-    if (!$result) {
-        print("Ошибка MySQL: " . mysqli_error($conn)); 
-    }
-    return mysqli_fetch_all($result, MYSQLI_ASSOC); 
-}
-
 /* функция формат цены */
 
 function makePriceFormat($price) {
