+++ index.php
@@ -1,30 +1,15 @@
 <?php
 
 require('inc/function.php'); // функции
+require('inc/queries.php'); // Запросы и подключение
 require('helpers.php'); // шаблонизатор
 
 $conn = getConn(); // Подключение к БД
 $categories = getCategories($conn); // Запрос Показать Таблицу Категории
-
-// Главная стр - Запрос показать активные лоты (врямя окончания не вышло), сортировать от последнего добавленного
-
-$sql = "SELECT items.*, categories.name AS category, symbol FROM items 
-    JOIN categories ON items.category_id = categories.id
-    WHERE items.ts_end > CURRENT_TIMESTAMP 
-    ORDER BY ts_add DESC 
-    LIMIT 9
-"; 
-
-$result = mysqli_query($conn, $sql);
-if (!$result) {
-    print("Ошибка MySQL: " . mysqli_error($conn)); 
-}
-$items = mysqli_fetch_all($result, MYSQLI_ASSOC);
-
-// Закрытие подключения к БД
+$items = getItems($conn); // Главная стр показать активные лоты, до 9 шт
 mysqli_close($conn);
 
-/* Шаблонизация - подключение подложики */
+/* Шаблонизация */
 
 $page_name = 'Главная';
