+++ templates/error.php
@@ -1,26 +1,13 @@
 
         <nav class="nav">
             <ul class="nav__list container">
-                <li class="nav__item">
-                    <a href="all-lots.html">Доски и лыжи</a>
-                </li>
-                <li class="nav__item">
-                    <a href="all-lots.html">Крепления</a>
-                </li>
-                <li class="nav__item">
-                    <a href="all-lots.html">Ботинки</a>
-                </li>
-                <li class="nav__item">
-                    <a href="all-lots.html">Одежда</a>
-                </li>
-                <li class="nav__item">
-                    <a href="all-lots.html">Инструменты</a>
-                </li>
-                <li class="nav__item">
-                    <a href="all-lots.html">Разное</a>
-                </li>
+                <?php 
+                /* Вкладывание простое горизонтальное меню, кроме главной страницы */
+                require(__DIR__ . '/../inc/mainMenuSimple.php'); 
+                ?>
             </ul>
         </nav>
         <section class="lot-item container">
-            <?= $error; ?>
+            <h2>404 Страница не найдена</h2>
+            <p>Данной страницы не существует на сайте.</p>
         </section>
